Implement game timer (#16)
* Initial commit

* Rename Timer

* Remove test code

* Finish Timer implementation

* Formatting

* Update TODOs

* Emit OnFinish event

Skipping makes no sense, just don't subscribe

* Add Timer resetter

* Rename Timer to avoid collision

* Add initial tests for Timer

This is most likely wrong, but it's just temporary anyway

* Update notes

* Run timer tests in parallel

* Rename event, add a skip, test the Stop

* Rename EventArgs

* Rewrite tests

Just do things locally to let it run in parallel, because we use timers, so we save extra seconds. I don't feel like mocking this to remove the timer :eyes:

* Update comment

* Implement initial timer on camera

Makes the camera speed increase after some time. This will be tweaked later.

* Update TODOs

* Update TODO

* Set default interval at 1

* Add lobby awaiting time

* Add a timer for removing the lobby ceiling

* Finish camera scene handling through timers

* Update docs

* Update naming/docs

* Change emit->raise

* Update TODOs

* Rename members

* Add `OnStart` event

* Expose lobby timer for read

* Use lobby timer from Arena scene

* Clamp provided internal, throw on non-multiple intervals

* Add a Timer component exposing EventTimer

This is later available through the Arena, which exposes this component and its timers from script

* Expose Timers from component

* Prevent jumpers from jumping through Timer

This is the post-game timer, where we wait a couple seconds before allowing the jumpers to jump again

* Add `PodiumTimeout`, update docs

* Work directly from Timers reference (arena)

* Update tests

* Get parent properly

* Get parent properly

* Rename timer

* Finish TODO